0.2.0-alpha.7: Quelling MSVC's concerns
Why is MSVC so picky? Nonetheless, here I tell it to calm down, and also cast some mismatching-sized types.

+++ include/internal/object_3d.h
@@ -25,7 +25,7 @@ struct Object3D_s
 {
     // The triangles must be contiguous in memory
     Triangle_t * triangles;
-    Uint32 numTriangles;
+    size_t numTriangles;
     vec3 pos;
     mat4 model;
     GLuint vbo;
